refactor(foundation): move ai_studio data types to foundation (A20)
Largest single extraction so far. Moves all pure-data request/response/
result types from `mockforge-core::ai_studio::*` to a new
`mockforge_foundation::ai_studio_types` module. The engines
(`ApiCritiqueEngine`, `SystemGenerator`, `BehavioralSimulator`,
`ArtifactFreezer`) stay in core because they hold `LlmClient` and
perform file I/O.

**Moved (~35 types across 4 source files):**

system_generator.rs:
- SystemGenerationRequest, GeneratedSystem, AppliedSystem,
  SystemArtifact, SystemMetadata

artifact_freezer.rs:
- FreezeRequest, FreezeMetadata, FrozenArtifact

api_critique.rs:
- CritiqueRequest, ApiCritique, AntiPattern, Redundancy, NamingIssue,
  ToneAnalysis, ToneIssue, RestructuringRecommendations,
  HierarchyImprovement, ConsolidationOpportunity,
  ResourceModelingSuggestion

behavioral_simulator.rs:
- CreateAgentRequest, NarrativeAgent, Intention (enum),
  BehavioralTraits, BehaviorPolicy, PolicyRule, AppState, CartState,
  CartItem, ErrorEncounter, Interaction, SimulateBehaviorRequest,
  SimulateBehaviorResponse, NextAction

Each core file now `pub use`s its relevant types from foundation for
backwards-compat. `mockforge-http/handlers/ai_studio.rs` migrated to
import data types directly from foundation; the 3 engine types stay
as `mockforge_core::ai_studio::*` imports.

Drops **69 deprecation warnings** from http (287 → 218). All 574 http
+ 1818 core + 42 foundation tests pass; workspace clippy clean.
