+++ crates/mockforge-core/src/ai_studio/api_critique.rs
@@ -39,222 +39,13 @@ use crate::intelligent_behavior::{
     types::LlmGenerationRequest,
 };
 use crate::Result;
-use serde::{Deserialize, Serialize};
+// Data types re-exported from foundation.
+pub use mockforge_foundation::ai_studio_types::{
+    AntiPattern, ApiCritique, ConsolidationOpportunity, CritiqueRequest, HierarchyImprovement,
+    NamingIssue, Redundancy, ResourceModelingSuggestion, RestructuringRecommendations,
+    ToneAnalysis, ToneIssue,
+};
 use serde_json::Value;
-
-/// Request for API critique analysis
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct CritiqueRequest {
-    /// API schema (OpenAPI JSON, GraphQL SDL, or Protobuf)
-    pub schema: Value,
-
-    /// Schema type: "openapi", "graphql", or "protobuf"
-    pub schema_type: String,
-
-    /// Optional focus areas for analysis
-    /// Valid values: "anti-patterns", "redundancy", "naming", "tone", "restructuring"
-    #[serde(default)]
-    pub focus_areas: Vec<String>,
-
-    /// Optional workspace ID for context
-    pub workspace_id: Option<String>,
-}
-
-/// API critique result
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct ApiCritique {
-    /// Detected anti-patterns
-    pub anti_patterns: Vec<AntiPattern>,
-
-    /// Detected redundancies
-    pub redundancies: Vec<Redundancy>,
-
-    /// Naming quality issues
-    pub naming_issues: Vec<NamingIssue>,
-
-    /// Emotional tone analysis
-    pub tone_analysis: ToneAnalysis,
-
-    /// Restructuring recommendations
-    pub restructuring: RestructuringRecommendations,
-
-    /// Overall score (0-100, higher is better)
-    pub overall_score: f64,
-
-    /// Summary of findings
-    pub summary: String,
-
-    /// Token usage for this critique
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub tokens_used: Option<u64>,
-
-    /// Estimated cost in USD
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub cost_usd: Option<f64>,
-}
-
-/// Detected anti-pattern in API design
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct AntiPattern {
-    /// Type of anti-pattern (e.g., "rest_violation", "inconsistent_naming", "poor_resource_modeling")
-    pub pattern_type: String,
-
-    /// Severity: "low", "medium", "high", "critical"
-    pub severity: String,
-
-    /// Location in schema (path, endpoint, etc.)
-    pub location: String,
-
-    /// Description of the issue
-    pub description: String,
-
-    /// Suggested fix
-    pub suggestion: String,
-
-    /// Example of the problem
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub example: Option<String>,
-}
-
-/// Detected redundancy in API
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct Redundancy {
-    /// Type of redundancy (e.g., "duplicate_endpoint", "overlapping_functionality")
-    pub redundancy_type: String,
-
-    /// Severity: "low", "medium", "high"
-    pub severity: String,
-
-    /// Affected endpoints/resources
-    pub affected_items: Vec<String>,
-
-    /// Description of the redundancy
-    pub description: String,
-
-    /// Suggested consolidation
-    pub suggestion: String,
-}
-
-/// Naming quality issue
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct NamingIssue {
-    /// Type of naming issue (e.g., "inconsistent_convention", "unclear_name", "abbreviation")
-    pub issue_type: String,
-
-    /// Severity: "low", "medium", "high"
-    pub severity: String,
-
-    /// Location (field name, endpoint name, etc.)
-    pub location: String,
-
-    /// Current name
-    pub current_name: String,
-
-    /// Description of the issue
-    pub description: String,
-
-    /// Suggested improvement
-    pub suggestion: String,
-}
-
-/// Emotional tone analysis
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct ToneAnalysis {
-    /// Overall tone assessment
-    pub overall_tone: String,
-
-    /// Issues found in error messages
-    pub error_message_issues: Vec<ToneIssue>,
-
-    /// Issues found in user-facing text
-    pub user_facing_issues: Vec<ToneIssue>,
-
-    /// Recommendations for improvement
-    pub recommendations: Vec<String>,
-}
-
-/// Tone issue in API text
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct ToneIssue {
-    /// Type of tone issue (e.g., "too_vague", "too_technical", "unfriendly")
-    pub issue_type: String,
-
-    /// Severity: "low", "medium", "high"
-    pub severity: String,
-
-    /// Location (error message, description, etc.)
-    pub location: String,
-
-    /// Current text
-    pub current_text: String,
-
-    /// Description of the issue
-    pub description: String,
-
-    /// Suggested improvement
-    pub suggestion: String,
-}
-
-/// Restructuring recommendations
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct RestructuringRecommendations {
-    /// Recommended resource hierarchy improvements
-    pub hierarchy_improvements: Vec<HierarchyImprovement>,
-
-    /// Consolidation opportunities
-    pub consolidation_opportunities: Vec<ConsolidationOpportunity>,
-
-    /// Resource modeling suggestions
-    pub resource_modeling: Vec<ResourceModelingSuggestion>,
-
-    /// Overall restructuring priority: "low", "medium", "high"
-    pub priority: String,
-}
-
-/// Hierarchy improvement suggestion
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct HierarchyImprovement {
-    /// Current structure
-    pub current: String,
-
-    /// Suggested structure
-    pub suggested: String,
-
-    /// Rationale
-    pub rationale: String,
-
-    /// Impact: "low", "medium", "high"
-    pub impact: String,
-}
-
-/// Consolidation opportunity
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct ConsolidationOpportunity {
-    /// Items that can be consolidated
-    pub items: Vec<String>,
-
-    /// Description of the opportunity
-    pub description: String,
-
-    /// Suggested consolidation approach
-    pub suggestion: String,
-
-    /// Benefits of consolidation
-    pub benefits: Vec<String>,
-}
-
-/// Resource modeling suggestion
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct ResourceModelingSuggestion {
-    /// Current modeling approach
-    pub current: String,
-
-    /// Suggested modeling approach
-    pub suggested: String,
-
-    /// Rationale
-    pub rationale: String,
-}
 
 /// API Critique Engine
 pub struct ApiCritiqueEngine {
