+++ crates/mockforge-core/src/ai_studio/behavioral_simulator.rs
@@ -40,282 +40,15 @@ use crate::intelligent_behavior::{
 };
 use crate::Result;
 use chrono::Utc;
-use serde::{Deserialize, Serialize};
+// Data types re-exported from foundation.
+pub use mockforge_foundation::ai_studio_types::{
+    AppState, BehaviorPolicy, BehavioralTraits, CartItem, CartState, CreateAgentRequest,
+    ErrorEncounter, Intention, Interaction, NarrativeAgent, NextAction, PolicyRule,
+    SimulateBehaviorRequest, SimulateBehaviorResponse,
+};
 use serde_json::Value;
 use std::collections::HashMap;
 use uuid::Uuid;
-
-/// Request to create a narrative agent
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct CreateAgentRequest {
-    /// Optional: Attach to existing persona ID
-    pub persona_id: Option<String>,
-
-    /// Optional: Behavior policy type (e.g., "bargain-hunter", "power-user", "churn-risk")
-    pub behavior_policy: Option<String>,
-
-    /// If true, generate new persona if persona_id is not provided or doesn't exist
-    pub generate_persona: bool,
-
-    /// Optional workspace ID
-    pub workspace_id: Option<String>,
-}
-
-/// Narrative agent that models user behavior
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct NarrativeAgent {
-    /// Agent ID
-    pub agent_id: String,
-
-    /// Persona ID (links to existing persona or new generated one)
-    pub persona_id: String,
-
-    /// Current intention
-    pub current_intention: Intention,
-
-    /// Session history of interactions
-    pub session_history: Vec<Interaction>,
-
-    /// Behavioral traits
-    pub behavioral_traits: BehavioralTraits,
-
-    /// Current app state awareness
-    pub state_awareness: AppState,
-
-    /// Behavior policy attached to persona
-    pub behavior_policy: BehaviorPolicy,
-
-    /// Created at timestamp
-    pub created_at: String,
-}
-
-/// User intention types
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-#[serde(rename_all = "snake_case")]
-pub enum Intention {
-    /// Explore products/content
-    Browse,
-
-    /// Actively looking to purchase
-    Shop,
-
-    /// Ready to complete purchase
-    Buy,
-
-    /// Leave due to frustration/error
-    Abandon,
-
-    /// Retry after error
-    Retry,
-
-    /// Move to different section
-    Navigate,
-
-    /// Search for something
-    Search,
-
-    /// Compare options
-    Compare,
-
-    /// Review/read content
-    Review,
-}
-
-/// Behavioral traits for the agent
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct BehavioralTraits {
-    /// Patience level (0.0-1.0, higher = more patient)
-    pub patience: f64,
-
-    /// Price sensitivity (0.0-1.0, higher = more price-sensitive)
-    pub price_sensitivity: f64,
-
-    /// Risk tolerance (0.0-1.0, higher = more risk-tolerant)
-    pub risk_tolerance: f64,
-
-    /// Technical proficiency (0.0-1.0, higher = more technical)
-    pub technical_proficiency: f64,
-
-    /// Engagement level (0.0-1.0, higher = more engaged)
-    pub engagement_level: f64,
-}
-
-/// Behavior policy attached to persona
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct BehaviorPolicy {
-    /// Policy type (e.g., "bargain-hunter", "power-user", "churn-risk")
-    pub policy_type: String,
-
-    /// Policy description
-    pub description: String,
-
-    /// Policy rules/behaviors
-    pub rules: Vec<PolicyRule>,
-}
-
-/// Policy rule for behavior
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct PolicyRule {
-    /// Condition that triggers this rule
-    pub condition: String,
-
-    /// Action to take
-    pub action: String,
-
-    /// Priority (higher = more important)
-    pub priority: i32,
-}
-
-/// App state awareness
-#[derive(Debug, Clone, Serialize, Deserialize, Default)]
-pub struct AppState {
-    /// Current page/section
-    pub current_page: Option<String>,
-
-    /// Cart state
-    pub cart: CartState,
-
-    /// Authentication state
-    pub authenticated: bool,
-
-    /// Recent errors encountered
-    pub recent_errors: Vec<ErrorEncounter>,
-
-    /// Current context
-    pub context: HashMap<String, Value>,
-}
-
-/// Cart state
-#[derive(Debug, Clone, Serialize, Deserialize, Default)]
-pub struct CartState {
-    /// Whether cart is empty
-    pub is_empty: bool,
-
-    /// Number of items
-    pub item_count: usize,
-
-    /// Total value
-    pub total_value: f64,
-
-    /// Items in cart
-    pub items: Vec<CartItem>,
-}
-
-/// Cart item
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct CartItem {
-    /// Item ID
-    pub item_id: String,
-
-    /// Item name
-    pub name: String,
-
-    /// Price
-    pub price: f64,
-
-    /// Quantity
-    pub quantity: usize,
-}
-
-/// Error encounter
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct ErrorEncounter {
-    /// Error type (e.g., "500", "timeout", "validation_error")
-    pub error_type: String,
-
-    /// Error message
-    pub message: String,
-
-    /// Timestamp
-    pub timestamp: String,
-
-    /// Number of times encountered
-    pub count: usize,
-}
-
-/// Interaction record
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct Interaction {
-    /// Interaction timestamp
-    pub timestamp: String,
-
-    /// Action taken
-    pub action: String,
-
-    /// Intention at time of action
-    pub intention: Intention,
-
-    /// Request details
-    pub request: Option<Value>,
-
-    /// Response details
-    pub response: Option<Value>,
-
-    /// Result (success, error, etc.)
-    pub result: String,
-}
-
-/// Request to simulate behavior
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct SimulateBehaviorRequest {
-    /// Optional: Use existing agent ID
-    pub agent_id: Option<String>,
-
-    /// Optional: Attach to existing persona
-    pub persona_id: Option<String>,
-
-    /// Current app state
-    pub current_state: AppState,
-
-    /// Trigger event (e.g., "error_500", "cart_empty", "payment_failed")
-    pub trigger_event: Option<String>,
-
-    /// Optional workspace ID
-    pub workspace_id: Option<String>,
-}
-
-/// Response from behavior simulation
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct SimulateBehaviorResponse {
-    /// Next action to take
-    pub next_action: NextAction,
-
-    /// New intention
-    pub intention: Intention,
-
-    /// Reasoning for the action
-    pub reasoning: String,
-
-    /// Updated agent state
-    pub agent: Option<NarrativeAgent>,
-
-    /// Token usage
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub tokens_used: Option<u64>,
-
-    /// Estimated cost
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub cost_usd: Option<f64>,
-}
-
-/// Next action to take
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct NextAction {
-    /// Action type (e.g., "GET", "POST", "navigate", "abandon")
-    pub action_type: String,
-
-    /// Target endpoint or page
-    pub target: String,
-
-    /// Optional request body
-    pub body: Option<Value>,
-
-    /// Optional query parameters
-    pub query_params: Option<HashMap<String, String>>,
-
-    /// Delay before action (ms)
-    pub delay_ms: Option<u64>,
-}
 
 /// Behavioral Simulator Engine
 pub struct BehavioralSimulator {
